% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{initialize_pareto}
\alias{initialize_pareto}
\title{Initialise a Pareto identification}
\usage{
initialize_pareto(design)
}
\arguments{
\item{design}{A \code{\link[=pareto_design]{pareto_design()}}.}
}
\value{
An object of class \code{paretoinfer_state}.
}
\description{
Initialise a Pareto identification
}
\examples{
design <- pareto_design(alternatives = 4, objectives = 2, bounds = c(0, 1), epsilon = 0.05)
state <- initialize_pareto(design)
state
}
