% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoinfer-package.R
\docType{package}
\name{paretoinfer-package}
\alias{paretoinfer}
\alias{paretoinfer-package}
\title{paretoinfer: Sequential Identification of Approximately Pareto-Optimal Alternatives under Noise}
\description{
Sequential elimination and identification of epsilon-Pareto-optimal alternatives among a finite set with noisy, correlated objectives. Simultaneous confidence sequences for objective differences, plausible Pareto sets, certified dominance and adaptive choice of the next evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/castlaboratory/paretoinfer}
  \item Report bugs at \url{https://github.com/castlaboratory/paretoinfer/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})

Authors:
\itemize{
  \item André Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})
}

}
\keyword{internal}
